refactor(index): follow our code style
file names should not be called index,
and should favour explicit names.

knowledge of the node module resolution
algorithm should not be required to work
on the dotcom-rendering project
