+++ dotcom-rendering/scripts/webpack/bundles.js
@@ -4,7 +4,7 @@
  * Set this to `true` if you want to serve a server-side experiment against
  * the a variant bundle A/B test.
  *
- * Ensure Sentry sampling in sentry/index.ts is adjusted for the sample
+ * Ensure Sentry sampling in sentry/sentryLoader.ts is adjusted for the sample
  * size of the test
  *
  * @type {boolean} prevent TS from narrowing this to its current value
