+++ dotcom-rendering/scripts/webpack/webpack.config.server.js
@@ -23,7 +23,7 @@ const swcLoader = [
 /** @type {(options: { sessionId: string } ) => import('webpack').Configuration} */
 module.exports = ({ sessionId }) => ({
 	entry: {
-		server: './src/server/index.ts',
+		server: './src/server/server.ts',
 	},
 	output: {
 		filename: `[name].js`,
