+++ dotcom-rendering/src/client/sentryLoader/sentryLoader.ts
